requirements.txt
infi.clickhouse-orm
psutil